+++ src/main/java/guckflix/crawlservice/entity/Movie.java
@@ -25,7 +25,7 @@ public class Movie {
 
     private float popularity;
 
-    private float voteCount;
+    private int voteCount;
 
     private float voteAverage;
